Type Alias storage_proofs_porep::stacked::StackedBucketGraph

pub type StackedBucketGraph<H> = StackedGraph<H, BucketGraph<H>>;

Aliased Type

struct StackedBucketGraph<H> { /* private fields */ }

Implementations

impl<H, G> StackedGraph<H, G>where H: Hasher, G: Graph<H> + ParameterSetMetadata + Sync + Send,

pub fn new( base_graph: Option<G>, nodes: usize, base_degree: usize, expansion_degree: usize, porep_id: PoRepID, api_version: ApiVersion ) -> Result<Self>

pub fn parent_cache(&self) -> Result<ParentCache>

Returns a reference to the parent cache.

pub fn copy_parents_data_exp( &self, node: u32, base_data: &[u8], exp_data: &[u8], hasher: Sha256, cache: Option<&mut ParentCache> ) -> Result<[u8; 32]>

pub fn copy_parents_data( &self, node: u32, base_data: &[u8], hasher: Sha256, cache: Option<&mut ParentCache> ) -> Result<[u8; 32]>

impl<H, G> StackedGraph<H, G>where H: Hasher, G: Graph<H> + ParameterSetMetadata + Sync + Send,

pub fn generate_expanded_parents( &self, node: usize, expanded_parents: &mut [u32] )

pub fn new_stacked( nodes: usize, base_degree: usize, expansion_degree: usize, porep_id: PoRepID, api_version: ApiVersion ) -> Result<Self>

pub fn base_graph(&self) -> &G

pub fn expansion_degree(&self) -> usize

pub fn base_parents(&self, node: usize, parents: &mut [u32]) -> Result<()>

pub fn expanded_parents(&self, node: usize, parents: &mut [u32]) -> Result<()>

Assign self.expansion_degree parents to node using an invertible permutation that is applied one way for the forward layers and one way for the reversed ones.